Rename x86_64 modules
Make the naming of the ARM and Intel-specific modules consistent by using the
Rust architecture name.
